requirements.txt
streamlit==1.42.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
supabase>=2.4.0
python-dotenv>=1.0.0
pandas>=2.2.0
streamlit-supabase-auth>=1.0.0
Pillow
pydantic>=2.0.0
beautifulsoup4
qrcode
